+++ src/ts/components/settings/subtitlesettings/CharacterEdgeColorSelectBox.ts
@@ -34,7 +34,7 @@ export class CharacterEdgeColorSelectBox extends SubtitleSettingSelectBox {
     this.addItem('yellow', i18n.getLocalizer('colors.yellow'));
     this.addItem('magenta', i18n.getLocalizer('colors.magenta'));
 
-    this.onItemSelected.subscribe((sender, key: string) => {
+    this.onItemSelectionChanged.subscribe((sender, key: string) => {
       this.settingsManager.characterEdgeColor.value = key;
 
       // Edge type and color go together, so we need to...