+++ src/ts/components/settings/subtitlesettings/FontSizeSelectBox.ts
@@ -66,7 +66,7 @@ export class FontSizeSelectBox extends SubtitleSettingSelectBox {
       this.selectItem(property.value);
     });
 
-    this.onItemSelected.subscribe((sender, key: string) => {
+    this.onItemSelectionChanged.subscribe((sender, key: string) => {
       this.settingsManager.fontSize.value = key;
     });